Move layout to session setup
Probably could have avoided all this if we just did it in main but not sure.
