+++ planner/app_hooks.py
@@ -5,7 +5,7 @@ from utils.logging_utils import setup_logger
 import logging
 from utils.geo_utils import process_geotiff
 from components.map import create_image_figure
-from components.planner import create_file_upload, create_data_col
+from components.planner import create_file_upload, create_data_col, add_image_tools
 
 tiff_file = "input/MADRID_RGB.tif"
 
@@ -16,7 +16,7 @@ def on_server_loaded(server_context):
 
     logger = setup_logger(name="waypoint_planner", log_level=logging.DEBUG)
     logger.info("Bokeh server has started!")
-    initialize_data(server_context, logger)
+    # initialize_data(server_context, logger) # Wrong, needs session context
     logger.info("Server startup completed.")
 
 
@@ -30,11 +30,9 @@ def initialize_data(server_context, logger):
     logger.debug("Processing initial GeoTIFF file.")
     rgba_image, bounds = process_geotiff(tiff_file, logger)
     image_source.data = {"image": [rgba_image], "bounds": [bounds]}
-    image_figure = create_image_figure(image_source)
 
     setattr(server_context, 'image_source', image_source)
     setattr(server_context, 'marker_source', marker_source)
-    # setattr(server_context, 'image_figure', image_figure)
 
     logger.debug(f"Server document (on_server_loaded): {curdoc()}")
     logger.info("Data initialized.")
@@ -44,48 +42,27 @@ def initialize_data(server_context, logger):
 def on_session_created(session_context):
     """Create a session-specific layout."""
 
-    # assert curdoc() == session_context.document, "Mismatch between curdoc() and session_context.document!"
-
-
     logger = setup_logger(name="waypoint_planner", log_level=logging.DEBUG)
-
-    logger.debug(f"Session document (on_session_created): {curdoc()}")
-    logger.debug(f"Session context: {session_context.id}")
-    # logger.debug(f"Session context: {curdoc().session_context.id}")
-    logger.debug(f"on_session_created curdoc(): {id(curdoc())}")
-    logger.debug(f"Current document: {curdoc()}")
-    logger.debug(f"Document roots before adding layout: {curdoc().roots}")
+    logger.debug(f'on_session_created: {id(session_context)}')
 
     # server_context = session_context.server_context
     initialize_data(session_context, logger)
-
     image_source = getattr(session_context, 'image_source')
     marker_source = getattr(session_context, 'marker_source')
 
-    logger.debug(f'image_source: {image_source}')
-
-    # Create a fresh image figure for this session
-    image_figure = create_image_figure(image_source)
-
     # Create the session-specific layout
-    # file_upload = FileInput(title="Select files:", accept=".tif,.tiff")
-    file_upload = create_file_upload()
+    image_figure = create_image_figure(image_source) # Create a fresh image figure for this session
+    file_upload = create_file_upload(image_source, image_figure, logger)
 
     # Define layout and add to the document
     image_container = column(file_upload, image_figure)
     image_container.sizing_mode = "stretch_both"
 
+    add_image_tools(image_figure, marker_source)
     data_col = create_data_col(image_figure, marker_source)
     planner_row = row(image_container, data_col)
     planner_row.sizing_mode = "stretch_both"
 
-    logger.debug(f"planner_row: {planner_row}")
-    logger.debug(f"image_container: {image_container}")
-    logger.debug(f"data_col: {data_col}")
+    setattr(session_context, 'planner_row', planner_row) # Pass to session, add to doc there
 
-
-    # session_context.document.add_root(planner_row)
-    # curdoc().add_root(planner_row)
-    setattr(session_context, 'planner_row', planner_row)
-
-    logger.debug(f"Document roots after adding layout: {curdoc().roots}")
+    logger.debug(f"on_session_created complete: {curdoc().roots}")
