+++ planner/components/planner.py
@@ -2,28 +2,79 @@ from bokeh.layouts import column, row
 from bokeh.models import (
     CrosshairTool, TableColumn, DataTable, CustomJS,
     ColumnDataSource, PointDrawTool, Button, Div,
-    FileInput, 
+    FileInput, Range1d 
 )
-from bokeh.document import without_document_lock
+from bokeh.plotting import curdoc
+from functools import partial
 
 from utils.geo_utils import process_geotiff, plan_traversal
 
-def create_file_upload():
+IMAGE_DOWNSAMPLE = 5 # Ratio by which size reduced 
+
+
+def create_file_upload(image_source, image_figure, logger):
     # FileInput widget
-    file_input = FileInput(accept=".tif,.tiff")
+    file_upload = FileInput(accept=".tif,.tiff")
+
+    async def process_and_update(file_contents):
+        rgba_image, bounds = process_geotiff(file_contents, logger, downsample_factor=IMAGE_DOWNSAMPLE)
+        image_source.data = {"image": [rgba_image], "bounds": [bounds]}
+        logger.debug(f"Updated figure bounds to: x_range=({bounds.left}, {bounds.right}), y_range=({bounds.bottom}, {bounds.top})")
+
+        bounds = image_source.data["bounds"][0]
+        # image_figure = getattr(SERVER_CONTEXT, 'image_figure')
+
+        # Get rid of possible previous image
+        image_figure.renderers = [
+            r for r in image_figure.renderers if not isinstance(r, type(image_figure.image_rgba))
+        ]
+
+        # Add a new renderer with the updated image
+        image_figure.image_rgba(
+            image="image",
+            source=image_source,
+            x=bounds.left,
+            y=bounds.bottom,
+            dw=bounds.right - bounds.left,
+            dh=bounds.top - bounds.bottom,
+        )
+
+        image_figure.update(
+            x_range = Range1d(bounds.left, bounds.right),
+            y_range = Range1d(bounds.bottom, bounds.top)
+        )
+        # image_figure.x_range = Range1d(bounds.left, bounds.right) # Update the figure bounds
+        # image_figure.y_range = Range1d(bounds.bottom, bounds.top)
+        
+        # Make sure points on top of map image
+        image_figure.renderers = image_figure.renderers[-1:] + image_figure.renderers[:-1]
+
+        logger.debug("Updated image figure")
+        logger.debug(f"x_range=({image_figure.x_range.start}, {image_figure.x_range.end}), y_range=({image_figure.y_range.start}, {image_figure.y_range.end})")
+
 
     # Callback for file upload
     def upload_callback(attr, old, new):
-        """Handle file upload."""
-        file_contents = file_input.value
-        if file_contents:
-            process_geotiff(file_contents)  # Remove header and decode
+        """
+        Handle image updates from uploaded files.
+        """
+        try:
+            file_contents = file_upload.value
+            if not file_contents:
+                logger.warning("No file contents uploaded!")
+                return
+            
+            logger.debug(f"Uploaded file size: {len(file_contents) / (1024 * 1024):.2f} MB")            
+            curdoc().add_next_tick_callback(partial(process_and_update, file_contents=file_contents))
+            logger.debug("Image processing completed")
+
+        except Exception as e:
+            logger.error(f"Error during file upload: {e}", exc_info=True)
 
     # Goes with file_input object
-    # Moved here for scoping bc using globals like a dumbass
-    file_input.on_change("value", upload_callback)
+    file_upload.on_change("value", upload_callback)
 
-    return file_input
+    return file_upload
 
 
 def add_image_tools(image_figure, marker_source):
