+++ planner/main.py
@@ -4,20 +4,8 @@ from bokeh.models import (
     FileInput, Range1d, CrosshairTool, PointDrawTool,
     Div, CustomJS,
 )
-from types import SimpleNamespace
-
-from utils.geo_utils import process_geotiff
 from utils.logging_utils import setup_logger
 import logging
-from components.planner import create_data_col, add_image_tools
-from components.map import create_image_figure
-
-# import cProfile
-
-from tornado.ioloop import IOLoop
-from concurrent.futures import ThreadPoolExecutor
-import asyncio
-from functools import partial
 
 IMAGE_DOWNSAMPLE = 5 # Ratio by which size reduced 
 
@@ -40,60 +28,60 @@ curdoc().add_root(planner_row)
 # Placeholder for bounds as a SimpleNamespace
 # default_bounds = SimpleNamespace(left=0, right=1000, bottom=0, top=1000)
 
-async def process_and_update(file_contents):
-    rgba_image, bounds = process_geotiff(file_contents, logger, downsample_factor=IMAGE_DOWNSAMPLE)
-    image_source.data = {"image": [rgba_image], "bounds": [bounds]}
-    logger.debug(f"Updated figure bounds to: x_range=({bounds.left}, {bounds.right}), y_range=({bounds.bottom}, {bounds.top})")
+# async def process_and_update(file_contents):
+#     rgba_image, bounds = process_geotiff(file_contents, logger, downsample_factor=IMAGE_DOWNSAMPLE)
+#     image_source.data = {"image": [rgba_image], "bounds": [bounds]}
+#     logger.debug(f"Updated figure bounds to: x_range=({bounds.left}, {bounds.right}), y_range=({bounds.bottom}, {bounds.top})")
 
-    bounds = image_source.data["bounds"][0]
-    # image_figure = getattr(SERVER_CONTEXT, 'image_figure')
+#     bounds = image_source.data["bounds"][0]
+#     # image_figure = getattr(SERVER_CONTEXT, 'image_figure')
 
-    # Get rid of possible previous image
-    image_figure.renderers = [
-        r for r in image_figure.renderers if not isinstance(r, type(image_figure.image_rgba))
-    ]
+#     # Get rid of possible previous image
+#     image_figure.renderers = [
+#         r for r in image_figure.renderers if not isinstance(r, type(image_figure.image_rgba))
+#     ]
 
-    # Add a new renderer with the updated image
-    image_figure.image_rgba(
-        image="image",
-        source=image_source,
-        x=bounds.left,
-        y=bounds.bottom,
-        dw=bounds.right - bounds.left,
-        dh=bounds.top - bounds.bottom,
-    )
+#     # Add a new renderer with the updated image
+#     image_figure.image_rgba(
+#         image="image",
+#         source=image_source,
+#         x=bounds.left,
+#         y=bounds.bottom,
+#         dw=bounds.right - bounds.left,
+#         dh=bounds.top - bounds.bottom,
+#     )
 
-    image_figure.update(
-        x_range = Range1d(bounds.left, bounds.right),
-        y_range = Range1d(bounds.bottom, bounds.top)
-    )
-    # image_figure.x_range = Range1d(bounds.left, bounds.right) # Update the figure bounds
-    # image_figure.y_range = Range1d(bounds.bottom, bounds.top)
+#     image_figure.update(
+#         x_range = Range1d(bounds.left, bounds.right),
+#         y_range = Range1d(bounds.bottom, bounds.top)
+#     )
+#     # image_figure.x_range = Range1d(bounds.left, bounds.right) # Update the figure bounds
+#     # image_figure.y_range = Range1d(bounds.bottom, bounds.top)
     
-    # Make sure points on top of map image
-    image_figure.renderers = image_figure.renderers[-1:] + image_figure.renderers[:-1]
+#     # Make sure points on top of map image
+#     image_figure.renderers = image_figure.renderers[-1:] + image_figure.renderers[:-1]
 
-    logger.debug("Updated image figure")
-    logger.debug(f"x_range=({image_figure.x_range.start}, {image_figure.x_range.end}), y_range=({image_figure.y_range.start}, {image_figure.y_range.end})")
+#     logger.debug("Updated image figure")
+#     logger.debug(f"x_range=({image_figure.x_range.start}, {image_figure.x_range.end}), y_range=({image_figure.y_range.start}, {image_figure.y_range.end})")
 
 
-# Callback for file upload
-def upload_callback(attr, old, new):
-    """
-    Handle image updates from uploaded files.
-    """
-    try:
-        file_contents = file_upload.value
-        if not file_contents:
-            logger.warning("No file contents uploaded!")
-            return
+# # Callback for file upload
+# def upload_callback(attr, old, new):
+#     """
+#     Handle image updates from uploaded files.
+#     """
+#     try:
+#         file_contents = file_upload.value
+#         if not file_contents:
+#             logger.warning("No file contents uploaded!")
+#             return
           
-        logger.debug(f"Uploaded file size: {len(file_contents) / (1024 * 1024):.2f} MB")            
-        curdoc().add_next_tick_callback(partial(process_and_update, file_contents=file_contents))
-        logger.debug("Image processing completed")
+#         logger.debug(f"Uploaded file size: {len(file_contents) / (1024 * 1024):.2f} MB")            
+#         curdoc().add_next_tick_callback(partial(process_and_update, file_contents=file_contents))
+#         logger.debug("Image processing completed")
 
-    except Exception as e:
-        logger.error(f"Error during file upload: {e}", exc_info=True)
+#     except Exception as e:
+#         logger.error(f"Error during file upload: {e}", exc_info=True)
 
 
 # Bokeh application layout
